refactor: routes file structure
