+++ src/App.vue
@@ -1,13 +1,9 @@
 <template>
-  <Header />
   <RouterView />
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import Header from '@/components/Header.vue'
 
-export default defineComponent({
-  components: { Header }
-})
+export default defineComponent({})
 </script>
