+++ tsconfig.node.json
@@ -5,5 +5,5 @@
     "moduleResolution": "Node",
     "allowSyntheticDefaultImports": true
   },
-  "include": ["../vite.config.ts"]
+  "include": ["./vite.config.ts"]
 }
